+++ fluent_blogs/templatetags/fluent_blogs_tags.py
@@ -4,7 +4,16 @@ from django.template import Library
 from fluent_blogs.models import get_entry_model
 from fluent_blogs.models.query import query_entries, query_tags
 from tag_parser import template_tag
-from tag_parser.basetags import BaseAssignmentOrInclusionNode
+from tag_parser.basetags import BaseAssignmentOrInclusionNode, BaseAssignmentOrOutputNode
+
+BlogPage = None
+
+HAS_APP_URLS = 'fluent_pages' in settings.INSTALLED_APPS
+if HAS_APP_URLS:
+    # HACK: accessing BlogPage directly. Apps are not completely separated this way.
+    # Should have some kind of registry and filter system (like middleware) instead.
+    from fluent_blogs.pagetypes.blogpage.models import BlogPage
+
 
 register = Library()
 
@@ -17,13 +26,45 @@ def blogurl(parser, token):
     For the former, URL resolving works via the normal '{% url "viewname" arg1 arg2 %}' syntax.
     For the latter, the URL resolving works via '{% appurl "viewname" arg1 arg2 %}' syntax.
     """
-    if 'fluent_pages' in settings.INSTALLED_APPS:
+    if HAS_APP_URLS:
         from fluent_pages.templatetags.appurl_tags import appurl
         return appurl(parser, token)
     else:
         # Using url from future, so the syntax is the same modern style.
         from django.templatetags.future import url
         return url(parser, token)
+
+
+@template_tag(register, 'get_entry_url')
+class GetEntryUrl(BaseAssignmentOrOutputNode):
+    """
+    Get the URL of a blog entry.
+
+    When using django-fluent-pages, this takes the current ``page`` variable into account.
+    It makes sure the blog entry is relative to the current page.
+
+    When django-fluent-pages is not used, using this is identical to calling ``entry.get_absolute_url()``.
+    """
+    min_args = 1
+    max_args = 1
+    takes_context = True
+
+    def get_value(self, context, *tag_args, **tag_kwargs):
+        entry = tag_args[0]
+
+        if HAS_APP_URLS:
+            # If the application supports mounting a BlogPage in the page tree,
+            # that can be used as relative start point of the entry.
+            page = context.get('page')
+            request = context.get('request')
+            if page is None and request is not None:
+                # HACK: access private django-fluent-pages var
+                page = getattr(request, '_current_fluent_page', None)
+
+            if page is not None and isinstance(page, BlogPage):
+                return page.get_entry_url(entry)
+
+        return entry.get_absolute_url()
 
 
 @register.filter
@@ -105,7 +146,7 @@ class GetEntriesNode(BlogAssignmentOrInclusionNode):
     )
     model = get_entry_model()
 
-    def get_value(self, *tag_args, **tag_kwargs):
+    def get_value(self, context, *tag_args, **tag_kwargs):
         # Query happens in the backend,
         # the templatetag is considered to be a frontend.
         qs = self.model.objects.all()
@@ -146,5 +187,5 @@ class GetPopularTagsNode(BlogAssignmentOrInclusionNode):
         'order', 'orderby', 'limit',
     )
 
-    def get_value(self, *tag_args, **tag_kwargs):
+    def get_value(self, context, *tag_args, **tag_kwargs):
         return query_tags(**tag_kwargs)